pyproject.toml
[tool.poetry]
name = "discord-tft-api"
version = "0.1.0"
description = ""
authors = ["Ryan Bui <ryan.bui@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-utils-traveler = "^0.0.5"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
